Waveform display and the shift of prediction to datapoint edior
